requirements.txt
blessings==1.6
bpython==0.15
certifi==2016.8.8
curtsies==0.2.6
foursquare==1!2015.5.26
future==0.15.2
greenlet==0.4.10
humanize==0.5.1
importlib==1.0.3
Pygments==2.1.3
python-telegram-bot==5.0.0
requests==2.11.1
six==1.10.0
urllib3==1.16
wcwidth==0.1.7
wheel==0.26.0
